refactor(tests): split selenium admin tests by model and add distribution autocomplete test
- Split test_f_admin.py into test_f_admin.py (generic), test_f_message.py, test_f_distribution.py
- Add Selenium test for DistributionList project/application autocomplete chain
